+++ tosixinch/data/_tosixinch.bash
@@ -5,11 +5,11 @@ _tosixinch()
     _init_completion -s || return
 
     case $prev in
-        --add-binary-extensions|--add-clean-attrs|--add-clean-tags|--css2|--elements-to-keep-attrs|--font-family|--font-mono|--font-sans|--font-scale|--font-serif|--font-size|--font-size-mono|--full-image|--guess|--landscape-size|--line-height|--pdfname|--portrait-size|--textindent|--textwidth|--toc-depth|--trimdirs|--user-agent|--viewcmd)
+        --add-binary-extensions|--add-clean-attrs|--add-clean-tags|--css2|--elements-to-keep-attrs|--font-family|--font-mono|--font-sans|--font-scale|--font-serif|--font-size|--font-size-mono|--full-image|--guess|--landscape-size|--line-height|--pdfname|--portrait-size|--selenium-chrome-path|--selenium-firefox-path|--textindent|--textwidth|--toc-depth|--trimdirs|--user-agent|--viewcmd)
             return
             ;;
         --browser-engine)
-            COMPREPLY=( $( compgen -W 'webkit webengine' -- "$cur" ) )
+            COMPREPLY=( $( compgen -W 'webkit webengine selenium-chrome selenium-firefox' -- "$cur" ) )
             return
             ;;
         --encoding)
@@ -48,7 +48,7 @@ _tosixinch()
 
     $split && return
 
-    COMPREPLY=( $( compgen -W '--add-binary-extensions --add-clean-attrs --add-clean-tags --appcheck --browser --browser-engine --check --cnvpath --convert --css2 --download --elements-to-keep-attrs --encoding --encoding-errors --extract --file --font-family --font-mono --font-sans --font-scale --font-serif --font-size --font-size-mono --force-download --ftype --full-image --guess --help --input --javascript --landscape-size --line-height --link --lxml --news --no-parts-download --nouserdir --orientation --parts-download --pdfname --portrait-size --prince --printout --quiet --raw --readability --readability-only --textindent --textwidth --toc --toc-depth --trimdirs --urllib --user-agent --userdir --verbose --version --view --viewcmd --weasyprint --wkhtmltopdf' -- "$cur" ) )
+    COMPREPLY=( $( compgen -W '--add-binary-extensions --add-clean-attrs --add-clean-tags --appcheck --browser --browser-engine --check --cnvpath --convert --css2 --download --elements-to-keep-attrs --encoding --encoding-errors --extract --file --font-family --font-mono --font-sans --font-scale --font-serif --font-size --font-size-mono --force-download --ftype --full-image --guess --help --input --javascript --landscape-size --line-height --link --lxml --news --no-parts-download --nouserdir --orientation --parts-download --pdfname --portrait-size --prince --printout --quiet --raw --readability --readability-only --selenium-chrome-path --selenium-firefox-path --textindent --textwidth --toc --toc-depth --trimdirs --urllib --user-agent --userdir --verbose --version --view --viewcmd --weasyprint --wkhtmltopdf' -- "$cur" ) )
     [[ $COMPREPLY == *= ]] && compopt -o nospace
     return